[TR-209] Move URL-linked data elements from select-report-type to route-data
