+++ public/canjs/controls/select-issue-type/select-issue-type.js
@@ -3,12 +3,13 @@ import { StacheElement, type, ObservableObject, ObservableArray, value, queues }
 import {updateUrlParam, pushStateObservable} from "../../routing/state-storage.js";
 import { bitoviTrainingIssueData } from "../../../examples/bitovi-training.js";
 
-import routeData, { issueHierarchyFromNormalizedIssues } from "../../routing/route-data.js";
+import routeData from "../../routing/route-data.js";
 import "../status-filter.js";
 
 import SimpleTooltip from "../../ui/simple-tooltip/simple-tooltip";
 
 import { DROPDOWN_LABEL } from "../../../shared/style-strings.js";
+import { issueHierarchyFromNormalizedIssues, toSelectedParts } from "../../routing/data-utils.js";
 
 const TOOLTIP = new SimpleTooltip();
 document.body.append(TOOLTIP);
@@ -33,13 +34,13 @@ document.body.append(RELEASES_TOOLTIP);
 class TypeSelectionDropdown extends StacheElement {
     static view = `
         {{# for(issueType of this.issueHierarchy) }}
-        <label class="px-4 py-2 block {{#eq(this.routeData.primaryIssueType, issueType.name)}}bg-blue-101{{else}}${hoverEffect}{{/eq}}"><input 
+        <label class="px-4 py-2 block {{#eq(this.primaryIssueType, issueType.name)}}bg-blue-101{{else}}${hoverEffect}{{/eq}}"><input 
             type="radio" 
             name="primaryIssueType" 
-            checked:from="eq(this.routeData.primaryIssueType, issueType.name)"
+            checked:from="eq(this.primaryIssueType, issueType.name)"
             on:change="this.onSelection(issueType.name)"/> {{issueType.name}}s </label>
         {{/ }}
-        <label class="px-4 py-2  block {{#eq(this.routeData.primaryIssueType, 'Release')}}bg-blue-101{{else}}${hoverEffect}{{/eq}} border-t border-t-2 border-t-neutral-301"
+        <label class="px-4 py-2  block {{#eq(this.primaryIssueType, 'Release')}}bg-blue-101{{else}}${hoverEffect}{{/eq}} border-t border-t-2 border-t-neutral-301"
             on:mouseenter="this.showReleases(scope.element)">
             Releases <img class="inline" src="/images/chevron-right-new.svg"/> 
         </label>
@@ -60,10 +61,10 @@ customElements.define("select-type-dropdown", TypeSelectionDropdown);
 class ReleasesTypeSelectionDropdown extends StacheElement {
     static view = `
         {{# for(issueType of this.issueHierarchy) }}
-        <label class="px-4 py-2 block {{#eq(this.routeData.secondaryIssueType, issueType.name)}}bg-blue-101{{else}}${hoverEffect}{{/eq}}"><input 
+        <label class="px-4 py-2 block {{#eq(this.secondaryIssueType, issueType.name)}}bg-blue-101{{else}}${hoverEffect}{{/eq}}"><input 
             type="radio" 
             name="primaryIssueType" 
-            checked:from="eq(this.routeData.secondaryIssueType, issueType.name)"
+            checked:from="eq(this.secondaryIssueType, issueType.name)"
             on:change="this.onSelection('Release', issueType.name)"/> {{issueType.name}}s </label>
         {{/ }}
     `
@@ -76,15 +77,15 @@ customElements.define("select-release-type-dropdown", ReleasesTypeSelectionDropd
 export class SelectIssueType extends StacheElement {
     static view = `
         <label for="reportOn" class="${DROPDOWN_LABEL}">Report on</label>
-        {{# not(this.routeData.primaryIssueType) }}
+        {{# not(this.primaryIssueType) }}
             <button class="rounded bg-neutral-201 px-3 py-1" id="reportOn">Loading ... </button>
         {{/ }}
-        {{# if(this.routeData.primaryIssueType) }}
+        {{# if(this.primaryIssueType) }}
             <button class="rounded bg-neutral-201 px-3 py-1 ${hoverEffect}" 
                 on:click="this.showChildOptions()" 
                 id="reportOn">
-                {{this.routeData.primaryIssueType}}s
-                {{# if(this.routeData.secondaryIssueType) }} / {{this.routeData.secondaryIssueType}}s {{/ if }}
+                {{this.primaryIssueType}}s
+                {{# if(this.secondaryIssueType) }} / {{this.secondaryIssueType}}s {{/ if }}
                 <img class="inline" src="/images/chevron-down.svg"/>
             </button>
         {{/ }}
@@ -99,6 +100,14 @@ export class SelectIssueType extends StacheElement {
                 this.simplifiedIssueHierarchy;
             
         },
+        get primaryIssueType() {
+            return this.routeData.selectedIssueType && toSelectedParts(this.routeData.selectedIssueType).primary;
+        },
+        get secondaryIssueType() {
+            return this.routeData.selectedIssueType && toSelectedParts(this.routeData.selectedIssueType).secondary;
+        }
+
+
     };
     onSelection(primaryType, secondaryType){
         if(secondaryType) {
@@ -111,8 +120,8 @@ export class SelectIssueType extends StacheElement {
     }
     showChildOptions(){
         let dropdown = new TypeSelectionDropdown().initialize({
-            primaryIssueType: this.routeData.primaryIssueType,
-            secondaryIssueType: this.routeData.secondaryIssueType,
+            primaryIssueType: this.primaryIssueType,
+            secondaryIssueType: this.secondaryIssueType,
             issueHierarchy: this.issueHierarchy,
             onSelection: this.onSelection.bind(this)
         })
