+++ public/canjs/controls/select-report-type/select-report-type.js
@@ -1,6 +1,8 @@
 import { StacheElement, type, ObservableObject, ObservableArray, value } from "../../../can.js";
+import { DROPDOWN_LABEL } from "../../../shared/style-strings.js";
 
 import {saveJSONToUrl,updateUrlParam} from "../../routing/state-storage.js";
+import routeData from "../../routing/route-data.js";
 
 import "../status-filter.js";
 
@@ -18,20 +20,6 @@ import SimpleTooltip from "../../ui/simple-tooltip/simple-tooltip";
 const TOOLTIP = new SimpleTooltip();
 document.body.append(TOOLTIP);
 
-const REPORTS = [{
-    key: "start-due",
-    name: "Gantt Chart"
-},{
-    key: "due",
-    name: "Scatter Plot"
-},{
-    key: "table",
-    name: "Estimation Table"
-},{
-    key: "group-grid",
-    name: "Group Grid"
-}];
-
 const hoverEffect = "hover:bg-neutral-301 cursor-pointer";
 
 class ReportSelectionDropdown extends StacheElement {
@@ -48,53 +36,41 @@ class ReportSelectionDropdown extends StacheElement {
 
 customElements.define("report-selection-dropdown", ReportSelectionDropdown);
 
-import { DROPDOWN_LABEL } from "../../../shared/style-strings.js";
-
 export class SelectReportType extends StacheElement {
     static view = `
         <label for="reportType" class="${DROPDOWN_LABEL}">Report type</label>
-        {{# not(this.primaryReportType) }}
+        {{# not(this.routeData.primaryReportType) }}
             ---
         {{/ }}
-        {{# if(this.primaryReportType) }}
+        {{# if(this.routeData.primaryReportType) }}
             <button 
                 class="rounded bg-neutral-201 px-3 py-1 ${hoverEffect}"
                 id="reportType"
                 on:click="this.showChildOptions()">{{this.primaryReportName}} <img class="inline" src="/images/chevron-down.svg"/></button>
         {{/ }}
     `;
-    static props ={
-        primaryReportType: saveJSONToUrl("primaryReportType", "start-due", String, {
-            parse: function(x) {
-                if( REPORTS.find( report => report.key === x) ) {
-                    return x;
-                } else {
-                    return "start-due"
-                }
-            }, 
-            stringify: x => ""+x
-        }),
-        reports: {
+    static props = {
+        routeData: {
             get default(){
-                return REPORTS;
+                return routeData;
             }
         },
         get primaryReportName(){
-            return this.reports.find( report => report.key === this.primaryReportType).name;
+            return this.routeData.reports.find( report => report.key === this.routeData.primaryReportType).name;
         }
     };
 
     showChildOptions(){
         let dropdown = new ReportSelectionDropdown().initialize({
-            primaryReportType: this.primaryReportType,
-            reports: this.reports,
+            primaryReportType: this.routeData.primaryReportType,
+            reports: this.routeData.reports,
             onSelection: this.onSelection.bind(this)
         })
 
         TOOLTIP.belowElementInScrollingContainer(this, dropdown);
     }
     onSelection(reportType) {
-        this.primaryReportType = reportType;
+        this.routeData.primaryReportType = reportType;
         TOOLTIP.leftElement();
     }
     connected(){
